examples/blog/pyproject.toml
[project]
name = "flaskr"
version = "1.0.0"
description = "The basic blog app built in the Flask tutorial."
readme = "README.md"
license = {file = "LICENSE.txt"}
maintainers = [{name = "Pallets", email = "contact@palletsprojects.com"}]
classifiers = ["Private :: Do Not Upload"]
dependencies = [
    "flask",
    "flask-alembic",
    "flask-sqlalchemy-lite",
]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = ["error"]

[tool.coverage.run]
source = ["flaskr", "tests"]
branch = true

[tool.mypy]
files = ["src/flaskr", "tests"]
show_error_codes = true
pretty = true
strict = true

[tool.pyright]
include = ["src/flaskr", "tests"]
typeCheckingMode = "basic"
